+++ sql/expression/comparison.go
@@ -481,229 +481,3 @@ var (
 	// and the elements of the right operand don't match.
 	ErrInvalidOperandColumns = errors.NewKind("operand should have %d columns, but has %d")
 )
-
-// In is a comparison that checks an expression is inside a list of expressions.
-type In struct {
-	comparison
-}
-
-// NewIn creates a In expression.
-func NewIn(left sql.Expression, right sql.Expression) *In {
-	return &In{newComparison(left, right)}
-}
-
-// Eval implements the Expression interface.
-func (in *In) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	typ := in.Left().Type().Promote()
-	leftElems := sql.NumColumns(typ)
-	left, err := in.Left().Eval(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if left == nil {
-		return nil, err
-	}
-
-	left, err = typ.Convert(left)
-	if err != nil {
-		return nil, err
-	}
-
-	switch right := in.Right().(type) {
-	case Tuple:
-		for _, el := range right {
-			if sql.NumColumns(el.Type()) != leftElems {
-				return nil, ErrInvalidOperandColumns.New(leftElems, sql.NumColumns(el.Type()))
-			}
-		}
-
-		for _, el := range right {
-			right, err := el.Eval(ctx, row)
-			if err != nil {
-				return nil, err
-			}
-
-			right, err = typ.Convert(right)
-			if err != nil {
-				return nil, err
-			}
-
-			cmp, err := typ.Compare(left, right)
-			if err != nil {
-				return nil, err
-			}
-
-			if cmp == 0 {
-				return true, nil
-			}
-		}
-
-		return false, nil
-	// TODO: fix this. circular dependency
-	// case *plan.Subquery:
-	// 	if leftElems > 1 {
-	// 		return nil, ErrInvalidOperandColumns.New(leftElems, 1)
-	// 	}
-	//
-	// 	typ := right.Type()
-	// 	values, err := right.EvalMultiple(ctx)
-	// 	if err != nil {
-	// 		return nil, err
-	// 	}
-	//
-	// 	for _, val := range values {
-	// 		val, err = typ.Convert(val)
-	// 		if err != nil {
-	// 			return nil, err
-	// 		}
-	//
-	// 		cmp, err := typ.Compare(left, val)
-	// 		if err != nil {
-	// 			return nil, err
-	// 		}
-	//
-	// 		if cmp == 0 {
-	// 			return true, nil
-	// 		}
-	// 	}
-	//
-	// 	return false, nil
-	default:
-		return nil, ErrUnsupportedInOperand.New(right)
-	}
-}
-
-// WithChildren implements the Expression interface.
-func (in *In) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	if len(children) != 2 {
-		return nil, sql.ErrInvalidChildrenNumber.New(in, len(children), 2)
-	}
-	return NewIn(children[0], children[1]), nil
-}
-
-func (in *In) String() string {
-	return fmt.Sprintf("%s IN %s", in.Left(), in.Right())
-}
-
-func (in *In) DebugString() string {
-	return fmt.Sprintf("%s IN %s", sql.DebugString(in.Left()), sql.DebugString(in.Right()))
-}
-
-// Children implements the Expression interface.
-func (in *In) Children() []sql.Expression {
-	return []sql.Expression{in.Left(), in.Right()}
-}
-
-// NotIn is a comparison that checks an expression is not inside a list of expressions.
-type NotIn struct {
-	comparison
-}
-
-// NewNotIn creates a In expression.
-func NewNotIn(left sql.Expression, right sql.Expression) *NotIn {
-	return &NotIn{newComparison(left, right)}
-}
-
-// Eval implements the Expression interface.
-func (in *NotIn) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	typ := in.Left().Type().Promote()
-	leftElems := sql.NumColumns(typ)
-	left, err := in.Left().Eval(ctx, row)
-	if err != nil {
-		return nil, err
-	}
-
-	if left == nil {
-		return nil, err
-	}
-
-	left, err = typ.Convert(left)
-	if err != nil {
-		return nil, err
-	}
-
-	switch right := in.Right().(type) {
-	case Tuple:
-		for _, el := range right {
-			if sql.NumColumns(el.Type()) != leftElems {
-				return nil, ErrInvalidOperandColumns.New(leftElems, sql.NumColumns(el.Type()))
-			}
-		}
-
-		for _, el := range right {
-			right, err := el.Eval(ctx, row)
-			if err != nil {
-				return nil, err
-			}
-
-			right, err = typ.Convert(right)
-			if err != nil {
-				return nil, err
-			}
-
-			cmp, err := typ.Compare(left, right)
-			if err != nil {
-				return nil, err
-			}
-
-			if cmp == 0 {
-				return false, nil
-			}
-		}
-
-		return true, nil
-	// TODO: fix this. circular dependency
-	// case *Subquery:
-	// 	if leftElems > 1 {
-	// 		return nil, ErrInvalidOperandColumns.New(leftElems, 1)
-	// 	}
-	//
-	// 	typ := right.Type()
-	// 	values, err := right.EvalMultiple(ctx)
-	// 	if err != nil {
-	// 		return nil, err
-	// 	}
-	//
-	// 	for _, val := range values {
-	// 		val, err = typ.Convert(val)
-	// 		if err != nil {
-	// 			return nil, err
-	// 		}
-	//
-	// 		cmp, err := typ.Compare(left, val)
-	// 		if err != nil {
-	// 			return nil, err
-	// 		}
-	//
-	// 		if cmp == 0 {
-	// 			return false, nil
-	// 		}
-	// 	}
-
-		return true, nil
-	default:
-		return nil, ErrUnsupportedInOperand.New(right)
-	}
-}
-
-// WithChildren implements the Expression interface.
-func (in *NotIn) WithChildren(children ...sql.Expression) (sql.Expression, error) {
-	if len(children) != 2 {
-		return nil, sql.ErrInvalidChildrenNumber.New(in, len(children), 2)
-	}
-	return NewNotIn(children[0], children[1]), nil
-}
-
-func (in *NotIn) String() string {
-	return fmt.Sprintf("%s NOT IN %s", in.Left(), in.Right())
-}
-
-func (in *NotIn) DebugString() string {
-	return fmt.Sprintf("%s NOT IN %s", sql.DebugString(in.Left()), sql.DebugString(in.Right()))
-}
-
-// Children implements the Expression interface.
-func (in *NotIn) Children() []sql.Expression {
-	return []sql.Expression{in.Left(), in.Right()}
-}
